+++ contracts/src/token/erc721/mod.rs
@@ -1,5 +1,5 @@
 //! Implementation of the [`Erc721`] token standard.
-use alloc::vec;
+use alloc::{vec, vec::Vec};
 
 use alloy_primitives::{uint, Address, FixedBytes, U128, U256};
 use openzeppelin_stylus_proc::interface_id;
@@ -17,7 +17,8 @@ use crate::utils::{
 };
 
 pub mod extensions;
-
+mod receiver;
+pub use receiver::IERC721Receiver;
 pub use sol::*;
 #[cfg_attr(coverage_nightly, coverage(off))]
 mod sol {
@@ -163,33 +164,6 @@ pub enum Error {
 impl MethodError for Error {
     fn encode(self) -> alloc::vec::Vec<u8> {
         self.into()
-    }
-}
-
-pub use receiver::IERC721Receiver;
-mod receiver {
-    #![allow(missing_docs)]
-    #![cfg_attr(coverage_nightly, coverage(off))]
-    stylus_sdk::stylus_proc::sol_interface! {
-        /// [`Erc721`] token receiver interface.
-        ///
-        /// Interface for any contract that wants to support `safe_transfers`
-        /// from [`Erc721`] asset contracts.
-        interface IERC721Receiver {
-            /// Whenever an [`Erc721`] `token_id` token is transferred
-            /// to this contract via [`Erc721::safe_transfer_from`].
-            ///
-            /// It must return its function selector to confirm the token transfer.
-            /// If any other value is returned or the interface is not implemented
-            /// by the recipient, the transfer will be reverted.
-            #[allow(missing_docs)]
-            function onERC721Received(
-                address operator,
-                address from,
-                uint256 token_id,
-                bytes calldata data
-            ) external returns (bytes4);
-        }
     }
 }